Update Map.cpp
The end functions are in case you want to generate a random maze and write it to a text file in the main function, taken from Shang-Chun's algorithm. That file would then have to be altered a bit and given to Map's LoadFromFile function.

GetItem still needs to be sorted out with the Item class. I'm trying to return a copy of gun, healthPotion, and sword based on Tile.h, but it's not working as expected.