Update Tile.cpp
Tile still needs to communicate with the Player and Enemy classes to Draw them - right now they take arguments that I'm not sure how to pass in. I think they are being changed to just have Player.Draw() and Enemy.Draw(), in which case the Draw() function here can be adjusted accordingly.

I'm having a tough time returning the gun, sword, or healthPotion as defined in Tile.h in the GetItem() function, but it seems necessary for the project. That needs to be fixed for final implementation.